refactor: consolidate MAC string utilities into mac_utils
pretty_mac/compact_mac lived in cloud_client.cpp's anonymous namespace
while upper_mac/parse_mac_pretty lived in ble_utils — four functions on
the same domain in two places, and the cloud client could not reuse the
ble_utils pair because that header drags NimBLE in. mac_utils.{h,cpp}
is pure string handling with no NimBLE/ESP-IDF dependency; ble_utils
keeps only the actual NimBLE glue.
