+++ components/switchbot_keypad_bridge/keypad_pairer.cpp
@@ -9,6 +9,7 @@
 #include "aes_ctr.h"
 #include "ble_utils.h"
 #include "keypad_advert.h"
+#include "mac_utils.h"
 
 namespace esphome {
 namespace switchbot_keypad_bridge {
