+++ components/switchbot_keypad_bridge/pairing_ui.cpp
@@ -10,6 +10,7 @@
 #include "esphome/core/log.h"
 #include "ble_utils.h"
 #include "keypad_advert.h"
+#include "mac_utils.h"
 
 namespace esphome {
 namespace switchbot_keypad_bridge {
